Coroutine refactor.
Finally getting rid of the `yieldreturn` library and replacing it with proper Kotlin coroutines. Also making it easier (hopefully) to reason about the behavior: each script/effect receives its own `MutableFrame`, which presents an immutable version of the result frame from the previous round and lets the scripti/effect write its own channel value modifications on top of it. After every script/effect is done with its frame, `ScriptManager` merges the results.

This led to one important change in semantics: it is no longer possible for a script to see a channel value that has been set in the same frame by an "earlier" script.

See `ScriptManager` and `ScriptBase.next()` for descriptions of how `ScriptManager` communicates with the individual scripts/effects.