Reimplemented the pgmplayer in python
The code should be a lot cleaner and easier to understand.
It talks directly to the OVDMX dongle and such that we no longer depend on OLA.
Python should be more than fast enough for this purpose, even when running on a Raspberry Pi.

The program relies on no third party packages, so any Python 3.5 and above should work.