Scripts for running the Docker container on Windows.
This was necessary because Docker bind mounts must be specified in a different
way on Windows. I didn't bother trying to figure out how to make a CMD
script do conditional stuff, so `start-dev.bat` is simpler than
`start-dev.sh` in that it only runs the container. I also added scripts
for separately creating a persistent container and running it, so that
we won't have to spend time on the Gradle download every time (though we
should perhaps do that in the Dockerfile instead).